# coursework/backend/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Krusty Crab CRM API
  description: |
    API для системы управления рестораном "Красти Крабс"
    
    Система предназначена для автоматизации бизнес-процессов ресторана:
    - Управление заказами и клиентами
    - Контроль меню и ингредиентов
    - Управление сотрудниками и рабочими сменами
    - Финансовый учет и отчётность
    - Программы лояльности
    - Мониторинг качества обслуживания
  version: 1.0.0
  contact:
    name: Krusty Crab Support
servers:
  - url: http://localhost:13228/api
    description: Local development server
...
paths:
...
  /auth/me:
    get:
      tags:
        - auth
      summary: Получить информацию о текущем пользователе
      description: Получение информации о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  username:
                    type: string
                  userType:
                    type: string
                    enum: [CLIENT, EMPLOYEE]
                  role:
                    type: string
                    nullable: true
                    description: Роль сотрудника (только для EMPLOYEE)
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== CLIENTS ==========
  # ========== CLIENTS ==========
...
  /clients/{clientId}:
    get:
      tags:
        - clients
      summary: Получить информацию о клиенте
      description: Получение информации о клиенте по ID
      operationId: getClientById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Информация о клиенте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - clients
      summary: Обновить данные клиента
      description: Обновление персональных данных клиента (FR-006)
      operationId: updateClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdateRequest'
      responses:
        '200':
          description: Данные клиента обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /orders/{orderId}:
    get:
      tags:
        - orders
      summary: Получить информацию о заказе
      description: Получение детальной информации о заказе (FR-003)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - orders
      summary: Обновить статус заказа
      description: Обновление статуса заказа (FR-008, FR-010, FR-013)
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Статус заказа обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /payments:
    get:
      tags:
        - payments
      summary: Получить список платежей
      description: История платежей с пагинацией (для клиента — только по своим заказам)
      operationId: getPayments
      security:
        - bearerAuth: []
      parameters:
        - name: success
          in: query
          description: Фильтр по результату платежа
          schema:
            type: boolean
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Максимум записей
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          description: Смещение (пагинация)
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: История платежей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'

    post:
      tags:
        - payments
      summary: Обработать платеж
      description: Обработка платежа за заказ (FR-019)
      operationId: processPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Платеж обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
  /payments/cash:
    post:
      tags:
        - payments
      summary: Обработать наличный платеж с расчетом сдачи
      description: Обработка наличного платежа с расчетом и отображением сдачи (FR-019)
      operationId: processCashPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashPaymentRequest'
      responses:
        '200':
          description: Платеж обработан, возвращена информация о сдаче
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /payments/order/{orderId}:
    get:
      tags:
        - payments
      summary: Получить информацию о платеже по заказу
      description: Получение информации о платеже для заказа
      operationId: getPaymentByOrderId
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Информация о платеже
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== REVIEWS ==========
  /reviews:
    post:
      tags:
        - reviews
      summary: Создать отзыв
      description: Создание отзыва о заказе (FR-004)
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Отзыв создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - reviews
      summary: Получить список отзывов
      description: Получение списка отзывов (FR-018, FR-020, для управляющего)
      operationId: getAllReviews
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          description: Фильтр по клиенту
          schema:
            type: integer
        - name: orderId
          in: query
          description: Фильтр по заказу
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
  /reviews/{reviewId}:
    get:
      tags:
        - reviews
      summary: Получить отзыв по ID
      description: Получение информации об отзыве
      operationId: getReviewById
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об отзыве
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== INVENTORY ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа. Получите токен через /auth/login/client или /auth/login/employee
  parameters:
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: integer
      description: ID клиента
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
      description: ID заказа
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        defaultAddress:
          type: string
        registeredAt:
          type: string
          format: date
        loyaltyPoints:
          type: integer
          minimum: 0
...
    ClientUpdateRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        defaultAddress:
          type: string
...
    Courier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleInfo:
          type: string
        available:
          type: boolean
        busy:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        createdByEmployeeId:
          type: integer
          nullable: true
        acceptedByEmployeeId:
          type: integer
          nullable: true
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        preparingAt:
          type: string
          format: date-time
          nullable: true
        readyAt:
          type: string
          format: date-time
          nullable: true
        cookingDurationSeconds:
          type: integer
          minimum: 0
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: decimal
        courierId:
          type: integer
          nullable: true
        courier:
          $ref: '#/components/schemas/Courier'
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
    OrderType:
      type: string
      enum:
        - dine_in
        - takeout
        - delivery
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - preparing
        - ready
        - delivering
        - delivered
        - completed
        - cancelled
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: number
          format: decimal
        paidAt:
          type: string
          format: date-time
          nullable: true
        success:
          type: boolean
    PaymentMethod:
      type: string
      enum:
        - cash
        - card
        - online
    PaymentRequest:
      type: object
      required:
        - orderId
        - method
      properties:
        orderId:
          type: integer
        method:
          $ref: '#/components/schemas/PaymentMethod'
        simulateFailure:
          type: boolean
          description: Только для метода online - принудительно смоделировать ошибку платежа (UC-2 альтернативный поток 6а)
          default: false
    CashPaymentRequest:
      type: object
      required:
        - orderId
        - amountReceived
      properties:
        orderId:
          type: integer
        amountReceived:
          type: number
          format: decimal
          minimum: 0
    ChangeResponse:
      type: object
      properties:
        orderTotal:
          type: number
          format: decimal
        amountReceived:
          type: number
          format: decimal
        change:
          type: number
          format: decimal
    Review:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        clientId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    ReviewCreateRequest:
      type: object
      required:
        - orderId
        - clientId
        - rating
      properties:
        orderId:
          type: integer
        clientId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string